+++ src/data/V1.ts
@@ -3,11 +3,9 @@ import {
   BigintIsh,
   Currency,
   CurrencyAmount,
-  currencyEquals,
   ETHER,
   JSBI,
   Pair,
-  Percent,
   Route,
   Token,
   TokenAmount,
@@ -156,32 +154,4 @@ export function useV1TradeExchangeAddress(trade: Trade | undefined): string | un
       : undefined
   }, [trade])
   return useV1ExchangeAddress(tokenAddress)
-}
-
-const ZERO_PERCENT = new Percent('0')
-const ONE_HUNDRED_PERCENT = new Percent('1')
-
-// returns whether tradeB is better than tradeA by at least a threshold percentage amount
-export function isTradeBetter(
-  tradeA: Trade | undefined,
-  tradeB: Trade | undefined,
-  minimumDelta: Percent = ZERO_PERCENT
-): boolean | undefined {
-  if (tradeA && !tradeB) return false
-  if (tradeB && !tradeA) return true
-  if (!tradeA || !tradeB) return undefined
-
-  if (
-    tradeA.tradeType !== tradeB.tradeType ||
-    !currencyEquals(tradeA.inputAmount.currency, tradeB.inputAmount.currency) ||
-    !currencyEquals(tradeB.outputAmount.currency, tradeB.outputAmount.currency)
-  ) {
-    throw new Error('Trades are not comparable')
-  }
-
-  if (minimumDelta.equalTo(ZERO_PERCENT)) {
-    return tradeA.executionPrice.lessThan(tradeB.executionPrice)
-  } else {
-    return tradeA.executionPrice.raw.multiply(minimumDelta.add(ONE_HUNDRED_PERCENT)).lessThan(tradeB.executionPrice)
-  }
 }
